chore: move @types/helpers to _helpers/types
